# Possible Improvements to `FixedContext`