As pointed out by @penelopeysm and @mhauru in #702 , `FixedContext` and `ConditionContext` doesn't quite do what we want for `.~` statements. One of the reasons we first introduced `fix` was to avoid hitting the `tilde_*_assume` pipeline to improve performance.

This PR implements a possible way of fixing #702 which involves overloading the `tilde_dot_assume` for `FixedContext` to handle cases where only parts of the LHS is `fixed`.

With this branch we can do stuff like fixing only a subset of a `.~` statement:
```julia
julia> @model function demo(n)
           m = Vector{Float64}(undef, n)
           m .~ Normal()
           s ~ Normal()

           return (; m)
       end
demo (generic function with 2 methods)

julia> model = fix(demo(2), @varname(m[1]) => 0.0);

julia> model()
(m = [0.0, -0.019050747797961672],)

julia> rand(model)
(var"m[2]" = 0.41118414073410653, s = -0.2798438193201507)
```

However, it requires overloading `tilde_dot_assume` for `FixedContext`, which does go slightly against what `tilde_*_assume` is meant to do (it's meant to be used for random variables, but clearly fixed variables are not random).

# Performance implications

IMO the interesting "case" is when we use `fix(::Model, ::NamedTuple)` since this is consistently what we consider as the "fast mode" in Turing.jl / DynamicPPL.jl, and we can always ask the user to provide the values as a `NamedTuple` if they really want performance.

There are a few different "approaches" we can take with `fixed` (and equally `condition`):
1. Use current approach. This calls `conditional_isfixed` + `getfixed_nested(__context__, vn)` in the main-body of a `@model`. _When it works_, this is very performant, as it's just compile-time generated check of `sym in names` for `VarName{sym}` and `NamedTuple{names}`.
2. Remove current approach in favour of overloading `tilde_*_assume` pipeline to extract the fixed values.
3. A "hybrid" approach. Use (1) whenever possible, _but_ if we hit the `tilde_*_assume`, we _also_ check there for `tilde_dot_assume` (so that we cover the cases listed in #702) by iterating over all the variables and defering to `tilde_assume` (i.e. without the `dot`).

I ran the following snippet on this branch and `#master`:

```julia
using Revise, DynamicPPL, Distributions, BenchmarkTools

@model function demo(n)
    m = Vector{Float64}(undef, n)
    m .~ Normal()
    s ~ Normal()
end

iter = 1:14
suite = BenchmarkGroup()
for nlog2 in iter
    n = 2^nlog2
    model = fix(demo(n), m=zeros(n))
    vi = VarInfo(model)
    suite[n] = @benchmarkable $model($vi)
end
results = run(suite, seconds=60)
OrderedDict(2^nlog2 => results[2^nlog2] for nlog2 in iter)
```

## On `#master` (Approach 1)
```julia
julia> OrderedDict(2^nlog2 => results[2^nlog2] for nlog2 in iter)
OrderedDict{Int64, BenchmarkTools.Trial} with 14 entries:
  2     => 125.000 ns
  4     => 125.000 ns
  8     => 125.000 ns
  16    => 125.000 ns
  32    => 84.000 ns
  64    => 125.000 ns
  128   => 83.000 ns
  256   => 166.000 ns
  512   => 208.000 ns
  1024  => 291.000 ns
  2048  => 333.000 ns
  4096  => 500.000 ns
  8192  => 1.000 μs
  16384 => 2.083 μs
```

## On this branch (Approach 3)
```julia
julia> OrderedDict(2^nlog2 => results[2^nlog2] for nlog2 in iter)
OrderedDict{Int64, BenchmarkTools.Trial} with 14 entries:
  2     => 166.000 ns
  4     => 166.000 ns
  8     => 208.000 ns
  16    => 208.000 ns
  32    => 208.000 ns
  64    => 166.000 ns
  128   => 125.000 ns
  256   => 250.000 ns
  512   => 291.000 ns
  1024  => 375.000 ns
  2048  => 375.000 ns
  4096  => 583.000 ns
  8192  => 1.000 μs
  16384 => 2.125 μs
```
As we can see, the performance difference is very, very minor. **However**, note that this PR still includes the `contetxual_isfixed` checki n the main body of the model.

## Replace current approach fully be overloading tilde

**If we remove this, i.e. only rely on overloading tilde-pipeline**, we get the following result:

```julia
julia> OrderedDict(2^nlog2 => results[2^nlog2] for nlog2 in iter)
OrderedDict{Int64, BenchmarkTools.Trial} with 14 entries:
  2     => 167.000 ns
  4     => 166.000 ns
  8     => 208.000 ns
  16    => 250.000 ns
  32    => 208.000 ns
  64    => 209.000 ns
  128   => 291.000 ns
  256   => 416.000 ns
  512   => 583.000 ns
  1024  => 875.000 ns
  2048  => 1.500 μs
  4096  => 2.416 μs
  8192  => 4.875 μs
  16384 => 9.125 μs
```

As we see here, once we have to rely on a for-loop over the variables to check, we do incur a "signfiicant" runtime overhead.

## Conclusion

Performing the check in `dot_tilde_assume` _only when explicitly needed_ doesn't really hurt performance much for `fix(::Model, ::NamedTuple)` (i.e. Approach 2 vs. Approach 1).

_However_, purely relying on Approach 3 (i.e. replacing current approach completely with overloading tilde assume) _does_ have quite a significant overhead for just evaluation (assuming this will be even worse when computing gradients).

Soooo I'm leaning towards Approach 2 (as is implemented in this branch), even though it does make things a bit uglier.