@torfjelde 2024-11-04
Thoughts on the incosistency of overriding `tilde_dot_assume` though? @penelopeysm @mhauru @yebai