@torfjelde 2024-11-04
> I don't know why but the varname does come out concretised in the chain though (and generated_quantities will use it)

Because we use `DynamicPPL.varname_and_value_leaves` in Turing.jl to convert it into `s[1, 1]` and `s[1, 2]`, etc.