@torfjelde 2024-11-04
Hmm, this is actually quite an annoying issue 😕 

It raises the question of whether `DynamicPPL.hasvalue` should support stuff like
```julia
DynamicPPL.hasvalue(OrderedDict(@varname(s[1,1]) => 0.0), @varname(s[:, 1]))
```
which I'm somewhat uncertain we want 😕

The current implementation assumes `vn` is a _strict_ "sub-key" of any of the keys present in `vals`

https://github.com/TuringLang/DynamicPPL.jl/blob/da6f9a015b4cabb6b5fec0696a66fdfa3e4cf71d/src/utils.jl#L872-L895

In an ideal world, this would also handle stuff like
```julia
DynamicPPL.hasvalue(OrderedDict(@varname(s[1,1]) => 0.0), @varname(s[:, 1]))
DynamicPPL.hasvalue(OrderedDict(@varname(s[:,1]) => [0.0]), @varname(s[1, 1]))
```
_but_ this will complicate the implementation of both `hasvalue` and `getvalue` quite a bit 😕