@mhauru 2024-11-04
This is a bit of a drive-by comment, but I've so far failed to wrap my head around how we use `VarName`s as keys in containers. I feel like sometimes we use `subsumes`, sometimes we don't, and maybe sometimes we also do some ad hoc half way thing. This is complicated further by having e.g. multiple backend storage structures for `(Simple)VarInfo`, and having getters and setters for them defined in a few different places (#654). Having to do string representations with Chains is also a big contributor to confusion.

I don't really have a proposal for how to change this, but for many cases like your above `hasvalue` examples it is not clear to me what I should expect to happen. A more systematic approach, with [mumblemumble interfacesomething design doc mumblemumble] might help solidify this.